requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.0
phonenumberslite==8.13.35
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
redis==5.0.3
sqlparse==0.5.0
typing_extensions==4.11.0
